First attempt at inheritance with passing tests
